+++ applovin_max/android/src/main/java/com/applovin/applovin_max/AppLovinMAXAdViewFactory.java
@@ -52,8 +52,10 @@ public class AppLovinMAXAdViewFactory
         // Optional params
         boolean isAutoRefreshEnabled = Boolean.TRUE.equals( params.get( "is_auto_refresh_enabled" ) ); // Defaults to true
         String placement = params.containsKey( "placement" ) ? (String) params.get( "placement" ) : null;
-        String customData = params.containsKey( "customData" ) ? (String) params.get( "customData" ) : null;
+        String customData = params.containsKey( "custom_data" ) ? (String) params.get( "custom_data" ) : null;
+        Map<String, Object> extraParameters = params.containsKey( "extra_parameters" ) ? (Map<String, Object>) params.get( "extra_parameters" ) : null;
+        Map<String, Object> localExtraParameters = params.containsKey( "local_extra_parameters" ) ? (Map<String, Object>) params.get( "local_extra_parameters" ) : null;
 
-        return new AppLovinMAXAdView( viewId, adUnitId, adFormat, isAutoRefreshEnabled, placement, customData, messenger, sdk, context );
+        return new AppLovinMAXAdView( viewId, adUnitId, adFormat, isAutoRefreshEnabled, placement, customData, extraParameters, localExtraParameters, messenger, sdk, context );
     }
 }